final change

+++ scr/src/eg/edu/alexu/csd/datastructure/calculator/cs74/MyCalculator.java
@@ -12,7 +12,7 @@ public class MyCalculator implements ICalculator  {
 
 	
 	public float divide(int x, int y) {
-		return x/y;
+		return (float)x/(float)y;
 	}
 
 }
